Move report generation into a class
Move the report generation into a `Report` class.  The ultimate goal
here is to generalize the report generation, but for now, just move the
contents into this class.  Several of the globals that were moved into
main need to be moved back out to be globals.  This can be remedied in a
future patch.

Signed-off-by: David Brown <david.brown@linaro.org>