CVE generation
The start of CVE generation.  Extract data from the JIRA ticket to
generate the CVE json data.  It is intended to be loaded into
https://vulnogram.github.io/ to be normalized, but should fill in many
of the fields.

It expects a block of text <CVE>...</CVE> in the issue description to
use as the description for the CVE.  The summary will be used directly.

Signed-off-by: David Brown <david.brown@linaro.org>