cvelate: Implement concurrent link lookup
The JIRA API requires the remotelink field to be retrieved via its own
API.  Because this needs to be queried per entity, serial API queries is
fairly slow.  Impelement a concurrent query that attempts to retrieve
all of these at once.

So far, I haven't seen any throttling from trying to retrieve all of
these at once, but maybe there should be a limit, and use a bounded
queue or something.

Signed-off-by: David Brown <david.brown@linaro.org>